% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup}
\alias{lookup}
\title{NOGA lookup table}
\format{
## `lookup`
A data frame with 1,790 rows and 10 columns:
\describe{
  \item{code}{NOGA code}
  \item{section}{section code}
  \item{division}{division code (2-digit)}
  \item{group}{group code (3-digit)}
  \item{class}{class code (4-digit)}
  \item{type}{type code (full 6-digit)}
  \item{name_en}{English label}
  \item{name_de}{German label}
  \item{name_fr}{French label}
  \item{name_it}{Italian label}
}
}
\source{
<https://www.i14y.admin.ch/it/catalog/datasets/HCL_NOGA/api>
}
\usage{
lookup
}
\description{
This data.frame contains all the NOGA codes and labels at various levels.
The labels are available in english, german, french and italian.
}
\keyword{datasets}
