% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noga_aggregate.R
\name{noga_aggregate}
\alias{noga_aggregate}
\title{Function to assign a lower noga level values a higher one
(e.g. assign divisions to noga groups) for aggregation.
Note that this function doesn't actually aggregate, it prepares your data
for aggregation. It also works only with noga-values not with its labels}
\usage{
noga_aggregate(var, language = "en", level, type = "numeric")
}
\arguments{
\item{var}{The variable containing the noga-codes or noga-values that you want
to aggregate. Must be numeric or string, factor variables are not supported.}

\item{language}{Required if either your input or output are labels.
One of "de" (German), "en" (English), "fr" (French) or "it" (Italian).
Defaults to English.}

\item{level}{For which higher NOGA-Level you want the values?
Must be one of "section" (1. level), "division" (2. level), "group" (3.),
"class" (4.). "Type" is not possible because it has no lower level to aggregate.}

\item{type}{Determines whether your output variable is "character" or
"numeric". Defaults to "numeric".}
}
\value{
A variable or vector that contains the values or labels of a higher
noga-unit for each lower noga-unit input.
}
\description{
Function to assign a lower noga level values a higher one
(e.g. assign divisions to noga groups) for aggregation.
Note that this function doesn't actually aggregate, it prepares your data
for aggregation. It also works only with noga-values not with its labels
}
\examples{
example.data <- c(479,433,990)
noga::noga_aggregate(var=example.data,language="en",level="section",type="")
}
