% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noga_levels}
\alias{noga_levels}
\title{NOGA code table}
\format{
## `noga_levels`
A tibble with 1,790 rows and 15 columns:
\describe{
  \item{section}{section code}
  \item{division}{division code (2-digit)}
  \item{group}{group code (3-digit)}
  \item{class}{class code (4-digit)}
  \item{type}{type code (full 6-digit)}
  \item{code}{NOGA code}
  \item{name_en}{English label}
  \item{name_de}{German label}
  \item{name_fr}{French label}
  \item{name_it}{Italian label}
  \item{division.n}{division code, numeric version}
  \item{group.n}{group code, numeric version}
  \item{class.n}{class code, numeric version}
  \item{type.n}{type code, numeric version}
  \item{code.n}{code, numeric version}
}
}
\source{
<https://www.i14y.admin.ch/it/catalog/datasets/HCL_NOGA/api>
}
\usage{
noga_levels
}
\description{
This tibble contains all the NOGA codes and labels at various levels.
The labels are available in english, german, french and italian.
}
\keyword{datasets}
