% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noga_recode.R
\name{noga_recode}
\alias{noga_recode}
\title{Function to recode a variable containing noga codes to labels and vice versa}
\usage{
noga_recode(var, language = "en", level = "auto", to = "auto")
}
\arguments{
\item{var}{The variable containing the noga-codes or noga-values that you want to recode. Must be numeric or string, factor variables are not supported.}

\item{language}{One of "de" (German), "en" (English), "fr" (French) or "it" (Italian). Defaults to English.}

\item{level}{The NOGA-Level you want to recode. Must be one of "auto" (Default), "section" (1. level), "division" (2. level), "group" (3.), "class" (4.) or "type" (5.). If set to "auto", the NOGA level will be identified based on the number of digits/characters of the maximum non-missing value of that variable.}

\item{to}{Recode values to labels or vice versa? Defaults to "auto" (will identify based on the input variable the recoding direction), or "values" will recode labels to values or "labels" will recode values to labels.}
}
\value{
a recoded variable in numeric() string format
}
\description{
Function to recode a variable containing noga codes to labels and vice versa
}
\examples{
example.data <-
noga::noga_recode(var=example.data$,language="fr",level="section",to="auto")
}
