% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_numeric_var.R
\name{transform_numeric_var}
\alias{transform_numeric_var}
\title{Function to transform numeric variables to avoid issue that leading 0's will be elimianted}
\usage{
transform_numeric_var(var)
}
\arguments{
\item{var}{The variable containing the noga-codes or noga-values that you}
}
\value{
the variable transformed so that it can be recoded
}
\description{
Function to transform numeric variables to avoid issue that leading 0's will be elimianted
}
